test(conftest): stub claude_agent_sdk at conftest collection — fix xdist race
The harness file at tests/harness/test_agent_runner_claude.py installs a
claude_agent_sdk stub at module-import time but guards it with
`if "claude_agent_sdk" in sys.modules: return`. Under pytest-xdist, if a
worker collects a sibling test module first that triggers
`from culture.clients.claude.daemon import AgentDaemon` (which transitively
imports the real SDK), the harness's stub-skip leaves later harness tests
bound to the real `ResultMessage` — which has stricter required positional
args than the harness stub expected.

This was a pre-existing bug, surfaced by Task 8 because that test also
loads daemon. Verified failing without my test under `-n auto` with 20
workers; CI's smaller worker count happened to avoid the racing collection
order, so the test matrix stayed green.

Fix: install the stub at the top of `tests/conftest.py`. pytest collects
conftest before any test module, so the stub lands in sys.modules first
and the per-file stub guards stay no-ops. Tests never actually call the
real SDK (everyone monkeypatches `query` / `send_request` etc., or sets
`skip_claude=True`), so swapping the SDK module out is transparent.

Also drops the now-redundant `_stub_claude_sdk_if_needed` helper from
tests/test_integration_agent_runner.py (added in this PR's first commit).

Verified: tests/ pass under `uv run pytest tests/ -n auto -q
--ignore=tests/test_cli_console_playwright.py` — 1145 passed in 37.74s.
